requirements.txt
pyspark==3.5.1
pandas>=2.2.0
psycopg2-binary>=2.9.9
streamlit>=1.36.0
python-dotenv>=1.0.1